+++ src/main/java/com/synrc/bert/Term.java
@@ -3,6 +3,7 @@ package com.synrc.bert;
 import fj.*;
 import fj.data.*;
 import static fj.data.Option.none;
+import java.math.BigDecimal;
 
 abstract class Term {
     private Term() {}
@@ -10,7 +11,7 @@ abstract class Term {
     public Res<String>  str() { return str(Res::ok).orSome(Res.fail(this + " is not a string"));}
     public Res<byte[]>  bin() { return bin(Res::ok).orSome(Res.fail(this + " is not a binary"));}
     public Res<Double>  float754() { return float754(Res::ok).orSome(Res.fail(this + " is not a float"));}
-    public Res<Double>  floatStr() { return floatStr(Res::ok).orSome(Res.fail(this + " is not a float"));}
+    public Res<BigDecimal>  floatStr() { return floatStr(Res::ok).orSome(Res.fail(this + " is not a float"));}
     public Res<Byte>    bt()  { return bt(Res::ok).orSome(Res.fail(this + " is not a byte"));}
     public Res<Integer> in()  { return in(Res::ok).orSome(Res.fail(this + " is not a integer"));}
 
@@ -18,7 +19,7 @@ abstract class Term {
     public static Term bin(byte[] bin) { return new Bin(bin); }
     public static Term list(List<Term> l) {return new Array(l.snoc(new Nil()));}
     public static Term float754(Double v) { return new Fload754(v);}
-    public static Term floatStr(Double v) { return new FloatStr(v);}
+    public static Term floatStr(BigDecimal v) { return new FloatStr(v);}
     public static Term bt(Byte v) { return new Bt(v.byteValue());}
     public static Term in(Integer v) { return new In(v.intValue());}
 
@@ -28,7 +29,7 @@ abstract class Term {
     public <T> Option<T> list(F<List<Term>, T> f) { return none(); }
     public <T> Option<T> nil(F0<T> f) { return none(); }
     public <T> Option<T> float754(F<Double,T> f) { return none(); }
-    public <T> Option<T> floatStr(F<Double,T> f) { return none(); }
+    public <T> Option<T> floatStr(F<BigDecimal,T> f) { return none(); }
     public <T> Option<T> bt(F<Byte,T> f) { return none(); }
     public <T> Option<T> in(F<Integer,T> f) { return none(); }
 
@@ -51,9 +52,9 @@ abstract class Term {
     }
 
     public static final class FloatStr extends Term {
-        final double v;
-        public FloatStr(double v) {  System.out.println("parsed " +v);this.v = v; }
-        public <T> Option<T> floatStr(F<Double,T> f) { return Option.some(f.f(v)); }
+        final BigDecimal v;
+        public FloatStr(BigDecimal v) {  System.out.println("parsed " +v);this.v = v; }
+        public <T> Option<T> floatStr(F<BigDecimal,T> f) { return Option.some(f.f(v)); }
     }
 
     public static final class Bin extends Term {
